refactor(models, dao): Rename `modelos` package to `models` and update references
- Standardized package naming by renaming `modelos` to `models`.
- Updated all class imports and `module-info.java` accordingly for consistency.
