+++ src/main/java/mx/edu/utez/warehousemanagerfx/controllers/AdminController.java
@@ -16,8 +16,8 @@ import javafx.scene.layout.BorderPane;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Pane;
 import javafx.stage.Stage;
-import mx.edu.utez.warehousemanagerfx.modelos.Warehouse;
-import mx.edu.utez.warehousemanagerfx.modelos.dao.WarehouseDao;
+import mx.edu.utez.warehousemanagerfx.models.Warehouse;
+import mx.edu.utez.warehousemanagerfx.models.dao.WarehouseDao;
 import mx.edu.utez.warehousemanagerfx.utils.RangeSliderAnimator;
 import mx.edu.utez.warehousemanagerfx.utils.routes.FXMLRoutes;
 import org.controlsfx.control.RangeSlider;
