+++ src/main/java/mx/edu/utez/warehousemanagerfx/models/dao/WarehouseDao.java
@@ -1,6 +1,6 @@
-package mx.edu.utez.warehousemanagerfx.modelos.dao;
+package mx.edu.utez.warehousemanagerfx.models.dao;
 
-import mx.edu.utez.warehousemanagerfx.modelos.Warehouse;
+import mx.edu.utez.warehousemanagerfx.models.Warehouse;
 import mx.edu.utez.warehousemanagerfx.utils.database.OracleDatabaseConnectionManager;
 
 import java.sql.*;
